requirements.txt
# SkyGuard - Raptor Alert System Dependencies
# Core AI/ML Libraries (choose one based on your preference)
# Option 1: PyTorch + YOLO (Recommended for SkyGuard)
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLO implementation

# Option 2: TensorFlow (Alternative - uncomment if preferred)
# tensorflow>=2.13.0
# tensorflow-object-detection-api

# Core Computer Vision
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# Hardware Integration
pyserial>=3.5
RPi.GPIO>=0.7.1  # For Raspberry Pi GPIO control
adafruit-circuitpython-neopixel>=6.3.0  # For LED indicators

# Camera and Video Processing
picamera2>=0.3.0  # For Raspberry Pi camera
imutils>=0.5.4

# Notification Services
requests>=2.31.0
twilio>=8.5.0  # For SMS notifications
pushbullet.py>=0.11.0  # For push notifications
pygame>=2.5.0  # For audio alerts

# Data Storage and Logging
sqlite3  # Built into Python
pandas>=2.0.0
matplotlib>=3.7.0

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
schedule>=1.2.0
psutil>=5.9.0

# Dataset Management
datasets>=3.1.0  # For downloading and managing training datasets

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Web Interface (Optional)
flask>=2.3.0
flask-cors>=4.0.0
